+++ requestor/service_manager/__init__.py
@@ -1,0 +1,1 @@
+from .yagna_erigon_manager import YagnaErigonManager
